requirements.txt
Babel==2.6.0
Django==2.1
django-phonenumber-field==2.0.1
phonenumbers==8.9.11
Pillow==5.2.0
pytz==2018.5
